removed class component version of image upload
