+++ src/components/RegisterForm/RegisterForm.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
-import ImageUpload from '../ImageUpload/ImageUploadFunctional';
+import ImageUpload from '../ImageUpload/ImageUpload';
 
 // Material-UI
 import {
